[cudamapper] Add functions for overlap rescue
This commit adds utility functions (incl. string_slice) that will
support similarity-based overlap rescue. The main new function is
rescue_overlap_ends, which takes a vector of overlaps, a query/target
FASTA parser, a number of bases to extend, and a minimum required
similarity. After calculating the similarity, the overlap is
extended if longer than <max_extend>. Currently this function is
only implemented for query heads.