[cudamapper] Implement chaining algorithm based on anchmers and [Guo et al](https://vast.cs.ucla.edu/sites/default/files/publications/minimap2-acc-approved.pdf).
Implements a transformation of minimap2's chaining algorithm similar to what
is used in Guo et al. This involves a forward search (up to N overlaps) and
a simple cumulative scoring algorithm. This is actually very similar to
the windowed chaining algorithm used for anchmers but does not degrade
when encountering repeated seeds (as anchmers and RLE do).